requirements.txt
tornado
derpconf
MySQL-python
